backend/expensy/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-oauth-toolkit==3.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
kombu==5.5.3
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
